requirements.txt
python-dotenv==1.0.1
streamlit==1.41.1
openai==1.59.7
duckduckgo_search==7.2.1
pyyaml==6.0.2
langsmith==0.2.11